backend/requirements.txt
aniso8601==9.0.1
attrs==23.2.0
bcrypt==4.1.2
blinker==1.7.0
click==8.1.7
Flask==3.0.2
Flask-JWT-Extended==4.6.0
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
importlib_resources==6.4.0
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
psycopg2==2.9.9
pycryptodome==3.20.0
PyJWT==2.8.0
python-decouple==3.8
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2024.1
referencing==0.34.0
rpds-py==0.18.0
SQLAlchemy==2.0.29
typing_extensions==4.10.0
Werkzeug==3.0.1
